Extract PanelTitle
